% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{add_attributes}
\alias{add_attributes}
\title{Add attribute metadata to a data.frame object}
\usage{
add_attributes(data, attributes)
}
\arguments{
\item{data}{(data.frame) Data object to add attributes information to.}

\item{attributes}{(data.frame) Data frame of attribute metadata to add.}
}
\value{
(data.frame())
}
\description{
Adds attribute metadata to a data.frame object.  Intended for use with a call
to 'EML:get_attributes' in order to extract metadata from an EML and attach
it a data.frame object as an attributes(data.frame) property.
}
\examples{
\dontrun{
data <- data.frame("depth" = c(1.2, 2.3), "temperature" = c(10.7, 9.5))
attributes <- data.frame("attributeName" = c("depth", "temperature"),
                         "attributeDefinition" = c("water depth in meters",
                                                    "temperature in celsius"),
                         "unit" = c("meter", "celsius"))
data <- add_attributes(data, attributes)

# View all attribute metadata
attributes(data)
# View attribute metadata for one variable
attributes(data)$depth

eml <- EML::read_eml(rawToChar(dataone::getObject(mnReal, "doi:10.18739/A2F299")))
attributes <- EML::get_attributes(eml@dataset@dataTable[[3]]@attributeList)[[1]]
data <- read.csv(
"https://arcticdata.io/metacat/d1/mn/v2/object/urn\%3Auuid\%3A11986ca1-5560-48ac-b8f9-0469ce561946")
data <- add_attributes(data, attributes)
}

}
\author{
Dominic Mullen \email{dmullen17@gmail.com}
}
