% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_creator_id.R
\name{add_creator_id}
\alias{add_creator_id}
\title{Add an ORCID or reference ID to a creator}
\usage{
add_creator_id(eml, orcid = NULL, id = NULL, surname = NULL)
}
\arguments{
\item{eml}{EML script to modify}

\item{orcid}{ORCID in the format 'https://orcid.org/WWWW-XXXX-YYYY-ZZZZ'}

\item{id}{reference ID. Character string to set reference ID for creators with additional roles (i.e. metadataProvider, etc.)}

\item{surname}{creator surname (last name), defaults to first creator if not specified. Not case-sensitive.}
}
\description{
This function allows you to add an ORCID or reference ID to a creator in EML.
}
\details{
The function invisibly returns the full EML, which
can be saved to a variable. It also prints the changed creator
entry so that it's easy to check that the appropriate change was
made. In addition to the EML, at least one of either the orcid or
reference id is required.

Note: updated creator information cannot be used as a
reference for associatedParties because the extra "role"
field is required. Also, the function does not (yet) account
for cases in which multiple creators have the same surname.
}
\examples{
\dontrun{
eml_path <- file.path(system.file(package = "datamgmt"), "dummy_meta_full.xml")
eml <- EML::read_eml(eml_path_original)
add_creator_id(eml, orcid = "https://orcid.org/WWWW-XXXX-YYYY-ZZZZ")

eml <- eml \%>\%
           add_creator_id(surname = "high-stakes",
                          orcid = "https://orcid.org/0000-1234-5678-4321",
                          id = "henrietta")

#use references to add updated contact info to Henrietta's other roles
eml@dataset@contact[[1]] <- new('contact', reference = "henrietta")
eml@dataset@metadataProvider[[1]] <- new('metadataProvider', reference = "henrietta")
}

}
\keyword{creator}
\keyword{eml}
\keyword{id}
\keyword{orcid}
