% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_party_orcid.R
\name{add_party_orcid}
\alias{add_party_orcid}
\title{add_party_orcid}
\usage{
add_party_orcid(eml, surName, givenName, orcid, replace = F)
}
\arguments{
\item{eml}{EML script to modify}

\item{surName}{surName of person to assign ORCID to}

\item{givenName}{givenName of person to assign ORCID to}

\item{orcid}{ORCID in the format 'https://orcid.org/WWWW-XXXX-YYYY-ZZZZ'}

\item{replace}{(TRUE/FALSE) If replace is set to FALSE, ORCIDs that already exist will not be overwritten}
}
\value{
Modified EML
}
\description{
This function allows you to add an ORCID to a ResponsibleParty in an EML
}
\keyword{EML}
\keyword{ORCID}
