% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{build_custom_units}
\alias{build_custom_units}
\title{Build shiny custom units table}
\usage{
build_custom_units(inputdf, standardUnits, inputdf2, unq_unitType, unq_parentSI)
}
\arguments{
\item{inputdf}{attributes table}

\item{standardUnits}{standardUnits table}

\item{inputdf2}{current units table}

\item{unq_unitType}{unique standardUnits unitTypes}

\item{unq_parentSI}{unique standardUnits parentSI}
}
\description{
Build shiny custom units table
}
