% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{build_factors}
\alias{build_factors}
\title{Build shiny factors table}
\usage{
build_factors(inputdf, inputdf2, data)
}
\arguments{
\item{inputdf}{attributes table}

\item{inputdf2}{current factors table}

\item{data}{initial data inputed by user}
}
\description{
Build shiny factors table
}
