% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{build_units_table}
\alias{build_units_table}
\title{Build units table}
\usage{
build_units_table(units)
}
\arguments{
\item{units}{(character) input units}
}
\description{
Get units and custom units
}
