% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_dataset.R
\name{categorize_dataset}
\alias{categorize_dataset}
\title{Interim method of categorizing datasets}
\usage{
categorize_dataset(doi, themes, coder, test = F, overwrite = F)
}
\arguments{
\item{doi}{(character) the doi formatted as doi:10.#####/#########}

\item{themes}{(list) themes of the dataset, can classify up to 5 - definition of the themes can be found here: https://docs.google.com/spreadsheets/d/1S_7iW0UBZLZoJBrHXTW5fbHH-NOuOb6xLghraPA4Kf4/edit#gid=1479370118}

\item{coder}{(character) your name, this is to identify who coded these themes}

\item{test}{(logical) for using the testing google sheet, defaults to FALSE}

\item{overwrite}{(logical) whether or not to update the themes}
}
\value{
NULL the result wil be written to an external google sheet
}
\description{
Please ask Jeanette or Jasmine to grant you access to the google sheet (https://docs.google.com/spreadsheets/d/1S_7iW0UBZLZoJBrHXTW5fbHH-NOuOb6xLghraPA4Kf4/edit#gid=1479370118)
}
\examples{
# categorize_dataset("doi:10.18739/A2QJ77Z09", c("biology", "oceanography"), "your name", T)
}
\author{
Jasmine Lai
}
