% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_data_objects}
\alias{clone_data_objects}
\title{Clone data pids between Dataone Member Nodes.}
\usage{
clone_data_objects(resource_map_pid, data_pids, from, to, n_max = 3L)
}
\arguments{
\item{resource_map_pid}{(character) Resource map for the data pids.}

\item{data_pids}{(character) Data object pids.}

\item{from}{(MNode) Dataone Member Node to clone data objects from}

\item{to}{(MNode) Dataone Member Node to clone data objects to}

\item{nmax}{(integer) Optional. Number of tries to download a DataOne object.
Can fail due to internet connectivity, suggested n_max >= 3.}
}
\value{
(list) Vector of data object pids.
}
\description{
This is a helper function for datamgmt::clone_package
}
