% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_one_package}
\alias{clone_one_package}
\title{Clone a Data Package between Member Nodes without its child packages.}
\usage{
clone_one_package(resource_map_pid, from, to)
}
\arguments{
\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}

\item{from}{(MNode) The Member Node to download from.}

\item{to}{(MNode) The Member Node to upload to.}
}
\value{
(list) List of all the identifiers in the new Data Package.
TODO switch remaining for loops to applys
TODO add more messages
}
\description{
The datamgmt wrapper function `clone_package` should be used instead.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
