% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_package}
\alias{clone_package}
\title{Clone a Data Package between Member Nodes.}
\usage{
clone_package(resource_map_pid, from, to, public = FALSE,
  clone_child_packages = FALSE)
}
\arguments{
\item{resource_map_pid}{(chraracter) The PID of the Resource Map of the package to clone between nodes.}

\item{from}{(MNode) The Member Node to download from.}

\item{to}{(MNode) The Member Node to upload to.}

\item{public}{(logical) Optional.  Set public read access.  Defaults to \code{FALSE}.}

\item{clone_child_packages}{(logical) Whether or not to clone the child packages.  Defaults to \code{FALSE}}
}
\description{
This function copies a Data Package from one DataOne member node to another.
This will not update the information in the metadata object.  This can also be used
to restore an older version of a Package to a member node, provided that the user subsequently
obsoletes the version of the package that they used to create the clone.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
