% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_packages}
\alias{clone_packages}
\title{Clone a Data Package between Member Nodes with its child packages.}
\usage{
clone_packages(resource_map_pid, from, to)
}
\arguments{
\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}

\item{from}{(MNode) The Member Node to download from.}

\item{to}{(MNode) The Member Node to upload to.}
}
\description{
This function copies a Data Package, including its child packages,
from one DataOne member node to another. It can also be used to copy an older
version of a Data Package to the same member node in order to restore it, provided
that the old Package is then obsoleted by the copied version.  This will not
update the pids in the metadata.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
