% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{compare_eml_to_package_pids}
\alias{compare_eml_to_package_pids}
\title{Return data object identifiers from 'dataTable' and 'otherEntity' objects
in an EML file}
\usage{
compare_eml_to_package_pids(eml, package_data_pids)
}
\arguments{
\item{eml}{(EML) EML object}
}
\value{
(list) List of all the identifiers in the EML.  Sorted into dataTable
and otherEntity identifiers
TODO better name for this function?
}
\description{
This function is a helper function for 'clone_package'.  It checks that all
of the data objects pids present in the metadata match those that the
resource map points to.  It also returns the data pids in the order in which
they appear in the EML: dataTable pids first and then otherEntity pids.  We
cannot update the new .xml file without this information.
}
