% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{convert_bytes}
\alias{convert_bytes}
\title{Format bytes to human readable format}
\usage{
convert_bytes(download_size)
}
\arguments{
\item{download_size}{(numeric) Total size in bytes

TODO - make this function more robust using gdata::humanReadable as a template}
}
\description{
This is a helper function for 'download_package'
}
