% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{create_attributes_table}
\alias{create_attributes_table}
\title{Allows editing of an attribute table and custom units table in a shiny environment}
\usage{
create_attributes_table(df, is.attribute.table = F,
  attributeDefinition = NULL, unit = NULL, measurementScale = NULL,
  domain = NULL, formatString = NULL, definition = NULL,
  numberType = NULL, missingValueCode = NULL,
  missingValueCodeExplanation = NULL)
}
\arguments{
\item{df}{A data frame of values}
}
\value{
Prints attribute table and custom units table to console
}
\description{
Allows editing of an attribute table and custom units table in a shiny environment
}
