% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_package2.R
\name{create_dummy_package2}
\alias{create_dummy_package2}
\title{Create dummy package}
\usage{
create_dummy_package2(mn, title = "A Dummy Package")
}
\arguments{
\item{mn}{(MNode) The member node to make the dummy package on}

\item{title}{(character) Optional. Title of package. Defaults to
"A Dummy Package".}
}
\description{
Creates a fuller package than \code{\link[arcticdatautils]{create_dummy_package}}
but is otherwise based on the same concept. This dummy
package includes multiple data objects (2 csv's, and 2 jpg's),
responsible parties, geographic locations, method steps, etc.
}
\author{
Irene Steves
}
