% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects_exist.R
\name{data_objects_exist}
\alias{data_objects_exist}
\title{Check if data objects exist in a list of Data Packages.}
\usage{
data_objects_exist(mn, pids, write_to_csv = FALSE, folder_path = NULL,
  file_name = NULL)
}
\arguments{
\item{mn}{(MNode/CNode) The Node to query for Object sizes}

\item{pids}{(character) The identifier of the Data Packages' Metadata}

\item{write_to_csv}{(logical) Optional. Write the query results to a csv}

\item{folder_path}{(character) Optional. Folder to write results to}

\item{file_name}{(character) Optional. Name of results file in csv format}
}
\value{
(data.frame) Data frame containing query results.
}
\description{
This function is primarily intended to assist Mark Schildhauer in preparation
for the Carbon synthesis working group.
}
\examples{
\dontrun{
cn <- CNode("PROD")
mn <- getMNode(cn, "urn:node:ARCTIC")
data_objects_exist(mn,
c("doi:10.5065/D60P0X4S", "urn:uuid:3ea5629f-a10e-47eb-b5ce-de10f8ef325b"))
}

}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
