% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{.onLoad}
\alias{.onLoad}
\title{Copies udunits2 xml files to datamgmt/UDUNITS
Updates uduntis2-accepted.xml with units from EML-units.xml}
\usage{
.onLoad(libname, pkgname)
}
\description{
Copies udunits2 xml files to datamgmt/UDUNITS
Updates uduntis2-accepted.xml with units from EML-units.xml
}
