% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{download_all_files}
\alias{download_all_files}
\title{Download multiple data objects}
\usage{
download_all_files(node, resource_map_pid, path)
}
\arguments{
\item{node}{(MNode) The Member Node to download data objects from.}

\item{resource_map_pid}{(character) The identifier of the data package's resource map.}

\item{path}{Path to write the shell script to.}
}
\value{
Outputs a shell script to the indicated path
}
\description{
This function outputs a shell script containing lines of wget commands. Executing them in the terminal will download all of the data objects in a data package to the local filesystem. This function is especially helpful when a data package contains too many data objects to download through the web interface.
}
\examples{
\dontrun{
# Set environment
cn <- dataone::CNode('PROD')
adc <- dataone::getMNode(cn,'urn:node:ARCTIC')

resource_map_pid <- 'resource_map_doi:10.18739/A2PG3B'
path = tempfile(fileext = '.sh')

download_all_files(adc, resource_map_pid, path)
expect_true(file.exists(path))
}

}
\author{
Rachel Sun rachelsun@ucsb.edu
}
