% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{download_data_objects}
\alias{download_data_objects}
\title{Download multiple data objects using their pids.}
\usage{
download_data_objects(mn, data_pids, out_paths, n_max = 3)
}
\arguments{
\item{mn}{(MNode) The Dataone Member Node to download the data objects from.}

\item{data_pids}{(character) A vector of Data object pids.}

\item{out_paths}{(character) A vector of file paths to download to.}

\item{n_max}{(numeric) Optional.  Number of attempts at downloading a Data object.}
}
\description{
Download mutiple dataone objects.  This is a helper function
for 'datamgmt::download_package'
}
