% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attributes.R
\name{download_eml_attributes}
\alias{download_eml_attributes}
\title{Download attribute (column) metadata from a Dataone Metadata object to csvs.}
\usage{
download_eml_attributes(eml, download_directory,
  prefix_file_names = FALSE)
}
\arguments{
\item{eml}{(S4) EML object.}

\item{download_directory}{(character) Directory to download attribute metadata csv's to.}

\item{prefix_file_names}{(logical) Optional.  Whether to prefix file names with the package metadata identifier.  This is useful when downloading files from multiple packages to one directory.}
}
\description{
Download attribute metadata from an EML object as csvs. The name
of each csv corresponds to the file name of the Data Object it describes.
This can be prepended with the package identifier by setting \code{prefix_file_names = TRUE} (recommended).
}
\examples{
\dontrun{
mn <- dataone::getMNode(cn, 'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(mn, "doi:10.18739/A23W02")))
attributes <- datamgmt::download_eml_attributes(eml, download_directory = tempdir(),
prefix_file_names = TRUE)
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
