% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{download_one_package}
\alias{download_one_package}
\title{Download one Package without its child Packages.}
\usage{
download_one_package(mn, resource_map_pid, download_directory,
  prefix_file_names = TRUE, download_column_metadata = FALSE,
  convert_excel_to_csv = FALSE)
}
\arguments{
\item{mn}{(MNode) The Member Node to download from.}

\item{resource_map_pid}{(character) The identifier of the Resource Map for the package to download.}

\item{download_directory}{(character) The path of the directory to download the package to.}

\item{prefix_file_names}{(logical) Optional.  Whether to prefix file names with the package metadata identifier.  This is useful when downloading files from multiple packages to one directory.}

\item{download_column_metadata}{(logical) Optional.  Whether to download attribute (column) metadata as csvs.  If using this its recommended to also set \code{prefix_file_names = TRUE}}

\item{convert_excel_to_csv}{(logical) Optional. Whether to convert excel files to csv(s).  This is not recommended if the separate csv files already exist in the package. The csv files are downloaded as sheetName_excelWorkbookName.csv}
}
\description{
This function downloads all of the Data Objects in a Data Package to the local filesystem.
It is particularly useful when a Data Package is too large to download using the web interface.
}
\examples{
\dontrun{
cn <- CNode("PROD")
mn <- getMNode(cn, "urn:node:ARCTIC")
download_one_package(mn, "resource_map_doi:10.18739/A2028W", "/home/dmullen")
}

}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
