% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_parties.R
\name{duplicated_parties}
\alias{duplicated_parties}
\title{duplicated_parties}
\usage{
duplicated_parties(eml)
}
\arguments{
\item{eml}{EML script to modify}
}
\value{
Modified EML
}
\description{
This function will replace all duplicated ResponsibleParty occurrences in an EML.
The function searches through all creators, contacts, metadataProviders, associatedParties,and publishers.
All unique surName, givenName occurances are recorded.
If a surName, givenName occurance is duplicated, every occurance after the first is replaced with the reference id of the first occurance.
}
\keyword{EML}
\keyword{ResponsibleParty}
\keyword{associatedParty}
\keyword{contact}
\keyword{creator}
\keyword{metadataProvider}
\keyword{publisher}
