% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{edit_attributes_table}
\alias{edit_attributes_table}
\title{Allows editing of an attribute table and custom units table in a shiny environment}
\usage{
edit_attributes_table(att_table)
}
\arguments{
\item{att_table}{An attribute table}
}
\value{
Prints attribute table and custom units table to console
}
\description{
Allows editing of an attribute table and custom units table in a shiny environment
}
