% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_csv.R
\name{excel_to_csv}
\alias{excel_to_csv}
\title{Convert excel workbook to multiple csv files}
\usage{
excel_to_csv(path, directory = NULL, ...)
}
\arguments{
\item{path}{(character) File location of the excel workbook.}

\item{directory}{(character) Optional.  Directory to download csv files to.}

\item{...}{Optional.  Allows arguments from \link[readxl]{read_excel}
Defaults to the base directory that \code{path} is located in.}
}
\value{
(invisible())
}
\description{
Converts an excel workbook into multiple csv files (one per tab).  Names the
files in the following format: sheetName_excelName.csv.
}
\author{
Dominic Mullen \email{dmullen17@gmail.com}
}
