% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{factor_att_table}
\alias{factor_att_table}
\title{Creates factors for attributes table}
\usage{
factor_att_table(inputdf)
}
\arguments{
\item{inputdf}{attributes table}
}
\description{
Creates factors for attributes table
}
