% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects_exist.R
\name{format_carbon_excel}
\alias{format_carbon_excel}
\title{Format carbon related Data Package Metadata for semantics analysis}
\usage{
format_carbon_excel(path, write_to_csv = FALSE, folder_path = NULL,
  file_name = NULL)
}
\arguments{
\item{path}{(character) Path to the local excel file}

\item{write_to_csv}{(logical) Optional. Write the query results to a csv}

\item{folder_path}{(character) Optional. Folder to write results to}

\item{file_name}{(character) Optional. Name of results file in csv format}
}
\value{
(data.frame) Data frame containing formatted and concatenated sheets.
}
\description{
This function is primarily intended to assist Mark Schildhauer's group in
preparation for the Carbon synthesis working group.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
