% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects_exist.R
\name{format_carbon_metadata}
\alias{format_carbon_metadata}
\title{Combine multiple excel sheets into one and write to local csv.}
\usage{
format_carbon_metadata(metadata, sheet_name)
}
\arguments{
\item{metadata}{(data.frame) Metadata data.frame() object to format.}

\item{sheet_name}{(character) Name of excel sheet}
}
\value{
(data.frame) Data frame containing formatted metadata.
}
\description{
This is a helper function. This function is primarily intended to assist Mark
Schildhauer's group in preparation for the Carbon synthesis working group.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
