% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{format_unit_split}
\alias{format_unit_split}
\title{Formats unit_split}
\usage{
format_unit_split(unit_split, form = "id", all_units = mem_load_all_units())
}
\arguments{
\item{unit_split}{(character) result of function get_unit_split}

\item{form}{('id', 'symbol', 'udunit', 'description'). 'id' is an EML id form. 'symbol' is an EML abbreviation form. 'udunit' is a udunits2 form.}

\item{all_units}{(data.frame)}
}
\value{
(character) formated unit_split
}
\description{
Formats unit_split
}
