% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_attributes.R
\name{get_all_attributes}
\alias{get_all_attributes}
\title{Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(LINK) Package}
\usage{
get_all_attributes(eml = NULL, link = NULL, node = NULL)
}
\arguments{
\item{eml}{(S4) Optional. The full eml document. Required if \code{link} is not specified.}

\item{link}{(character) Optional. The link to the Dataone Package. Required if \code{eml} is not specified.}

\item{node}{(MNode/CNode) Optional.  Defaults to Arctic Data Center Production node.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(LINK) Package
}
\examples{
\dontrun{
eml_pid <- "doi:10.18739/A23W02"
eml <- read_eml(rawToChar(getObject(mn, eml_pid)))
}
}
\author{
Dominic Mullen
}
