% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_csv}
\alias{get_csv}
\title{Get a csv file from a DataOne node}
\usage{
get_csv(mn, pid)
}
\arguments{
\item{mn}{A member node}

\item{pid}{The identifier of the data object}
}
\description{
Get a csv file from a DataOne node; allows you to access the data in your
scripts without downloading the data onto your hard drive. Ensures that the original
data stays intact. This function is a wrapper for \link[dataone]{getObject}.
}
\examples{
\dontrun{
cn <- CNode('PROD')
mn <- getMNode(cn,'urn:node:ARCTIC')

data <- get_csv(mn, "urn:uuid:087bf762-ea3d-444f-ba59-1c0bc02fd415")
}

}
\author{
Irene Steves
}
