% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_units.R
\name{get_custom_units}
\alias{get_custom_units}
\title{Get custom unit data frame}
\usage{
get_custom_units(units, quiet = FALSE)
}
\arguments{
\item{units}{(character) unit or vector of units}

\item{quiet}{(logical) if true will quiet console text}
}
\value{
(data.frame) custom unit data frame (will return a row of NAs if a unit cannot be formated in an EML form)
}
\description{
Uses the udunits2 unit library to format inputted unit into an EML unit form.
}
\examples{
#The following all return the same data frame.
get_custom_units('kilometersPerSquareSecond') #preferred input form
get_custom_units('Kilometers per seconds squared')
get_custom_units('km/s^2')
get_custom_units('km s-2')
get_custom_units('s-2 /     kilometers-1') #works but is not advised
}
