% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects_exist.R
\name{get_data_pids}
\alias{get_data_pids}
\title{Return the Data object identifiers associated with a Resource Map.}
\usage{
get_data_pids(mn, pid)
}
\arguments{
\item{mn}{(MNode/CNode) The Node to query}

\item{pid}{(character) The resource map identifier to query}
}
\value{
(character)
}
\description{
Return the Data object identifiers associated with a Resource Map.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
