% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_eml}
\alias{get_eml}
\title{Get an EML file from a DataOne node}
\usage{
get_eml(mn, pid)
}
\arguments{
\item{mn}{A member node}

\item{pid}{The identifier of the data object}
}
\description{
Get an EML file from a DataOne node. This function is a wrapper for
\link[dataone]{getObject}.
}
\examples{
\dontrun{
cn <- CNode('PROD')
mn <- getMNode(cn,'urn:node:ARCTIC')

eml <- get_eml(mn, "doi:10.18739/A2TP3N")
}


}
