% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attributes.R
\name{get_eml_attributes}
\alias{get_eml_attributes}
\title{Return attribute (column) metadata from a Dataone Metadata object.}
\usage{
get_eml_attributes(eml)
}
\arguments{
\item{metadata}{(S4) EML object.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Return attribute metadata from an EML object. This is largely a
wrapper for the function \code{get_attributes} from the EML Package
\url{https://github.com/ropensci/EML}.
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
mn <- dataone::getMNode(cn, 'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(mn, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_eml_attributes(mn, eml)

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_eml_attributes(knb, "doi:10.5063/F1639MWV")
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
