% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attributes.R
\name{get_eml_attributes_url}
\alias{get_eml_attributes_url}
\title{Return attribute (column) metadata from a DataONE Package URL.}
\usage{
get_eml_attributes_url(mn, url_path, write_to_csv = FALSE,
  prefix_file_names = FALSE, download_directory = NULL)
}
\arguments{
\item{mn}{(MNode/CNode) The DataONE Node that stores the Metadata object, from \url{https://cn.dataone.org/cn/v2/node}}

\item{url_path}{(character) The url of the DataONE Package.}

\item{write_to_csv}{(logical) Optional. Option whether to download the attribute metadata to csv's.  Defaults to \code{FALSE}}

\item{prefix_file_names}{(logical) Optional.  Whether to prefix file names with the package metadata identifier.
This is useful when downloading files from multiple packages to one directory.}

\item{download_directory}{(character) Optional.  Directory to download attribute metadata csv's to.
Required if \code{write_to_csv} is \code{TRUE}}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Return attribute metadata from an EML object or DataONE Package URL.
This is largely a wrapper for the function \code{get_attributes} from the EML Package
\url{https://github.com/ropensci/EML}.
}
\examples{
\dontrun{
attributes <- get_eml_attributes(mn,
"https://arcticdata.io/catalog/#view/doi:10.18739/A23W02")

# Download attribute metadata in csv format:
attributes <- get_eml_attributes(mn,
"https://arcticdata.io/catalog/#view/doi:10.18739/A23W02",
write_to_csv = TRUE,
download_directory = tempdir())
# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_eml_attributes(knb,
"https://knb.ecoinformatics.org/#view/doi:10.5063/F1639MWV")
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
