% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{get_eml_pids}
\alias{get_eml_pids}
\title{Return data identifiers from within an EML}
\usage{
get_eml_pids(eml)
}
\arguments{
\item{eml}{() an EML object}
}
\value{
Returns a list of data object pids in the eml
}
\description{
This function returns data object identifers present within an EML (electronic
metadata language) document.
}
