% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{get_format_type}
\alias{get_format_type}
\title{Return the 'formatType' of a Dataone object.}
\usage{
get_format_type(mn, pid)
}
\arguments{
\item{mn}{(MNode/CNode) The Node to query}

\item{pid}{(character) The unique object identifier}
}
\value{
(character)
}
\description{
Return the 'formatType' of a Dataone object.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
