% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_attributes.R
\name{get_meta_attributes}
\alias{get_meta_attributes}
\title{Return attribute metadata from an EML object or Dataone Package URL.}
\usage{
get_meta_attributes(node = "ADC", metadata)
}
\arguments{
\item{node}{(MNode/CNode) The Dataone Node that stores the Metadata object, from \url{https://cn.dataone.org/cn/v2/node}}

\item{metadata}{(S4 / character) Optional. Either the full eml S4 object document or the url of the Dataone Package.
The S4 object input is a more reliable method.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Return attribute metadata from an EML object or Dataone Package URL.
This is largely a wrapper for the function \code{get_attributes} from the EML Package
\url{https://github.com/ropensci/EML}.
}
\examples{
\dontrun{
node <- dataone::getMNode(cn, 'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(node, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_meta_attributes(metadata = eml)

attributes <- datamgmt::get_meta_attributes("ADC", https://arcticdata.io/catalog/#view/doi:10.18739/A23W02")

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_meta_attributes("https://knb.ecoinformatics.org/#view/doi:10.5063/F1639MWV")
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
