% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{get_object_metadata}
\alias{get_object_metadata}
\title{Return the object metadata of each data object in a Data Package}
\usage{
get_object_metadata(node, package_identifier, formatType = "DATA")
}
\arguments{
\item{node}{(MNode/CNode) The Node to query for Object sizes}

\item{formatType}{(character) Optional. Filter to just Objects of the given
formatType. One of METADATA, RESOURCE, or DATA or * for all types}

\item{resource_map_pid}{(character) The identifier of the Data Package's Resource Map}
}
\value{
(character) The formatId, fileName, and identifier of each data object in a package.
}
\description{
This function returns the formatId, fileName and identifier of each data
object in a Data Package.  This is a helper function for the function
'clone_package' - which copies a dataOne Data Package from one member node to
another.
}
