% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{get_parentSI_df}
\alias{get_parentSI_df}
\title{Get the unitType of unit_split}
\usage{
get_parentSI_df(udunit, all_units = mem_load_all_units(),
  EML_units = mem_load_EML_units())
}
\arguments{
\item{udunit}{(character) result of format_unit_split(unit_split, form = 'udunit')}

\item{all_units}{(data.frame)}

\item{EML_units}{(list)}
}
\value{
(data.frame) parentSI_df
}
\description{
Get the unitType of unit_split
}
