% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_units.R
\name{get_plural}
\alias{get_plural}
\title{Gets plural form of unit
Based on ut_form_plural function in xml.c of udunits2}
\usage{
get_plural(unit, udunits_units = mem_load_all_units())
}
\arguments{
\item{unit}{(character) unit}

\item{udunits_units}{(data.frame) load_udunits()}
}
\value{
(character) plural form of unit
}
\description{
Gets plural form of unit
Based on ut_form_plural function in xml.c of udunits2
}
