% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{get_resource_maps}
\alias{get_resource_maps}
\title{Return the Resource Map identifieres associated with a list of Dataone objects.}
\usage{
get_resource_maps(mn, pids)
}
\arguments{
\item{mn}{(MNode/CNode) The Node to query}

\item{pid}{(character) The unique object identifier}
}
\value{
(character)
}
\description{
Return the Resource Map identifieres associated with a list of Dataone objects.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
