% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_rt_csv}
\alias{get_rt_csv}
\title{Get an RT csv attachment}
\usage{
get_rt_csv(ticket, download_url)
}
\arguments{
\item{ticket}{(character) RT ticket number}

\item{download_url}{(character) RT download URL for the attachment}
}
\description{
This function allows you to read a csv attachment to an RT message
directly in R. It requires the RT ticket number and download URL
as arguments. The download URL can be accessed by right-clicking on
the link to download the attachment, and clicking "Copy Link Address."
}
\examples{
\dontrun{
data <- download_rt_csv(
    ticket = "13843",
    download_url = "https://support.nceas.ucsb.edu/rt/Ticket/Attachment/291896/209025/2016\%20ITEX\%20Barrow\%20Atqasuk\%20Climate\%20v1.csv")
}

}
\author{
Irene Steves
}
