% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_all_versions.R
\name{get_solr_fields}
\alias{get_solr_fields}
\title{Get all Solr fields}
\usage{
get_solr_fields()
}
\value{
(character) List of Solr fields
}
\description{
Simple helper function for 'query_all_versions' function to retrieve all Solr fields
}
\references{
Written by Irene in the reference guide at https://github.com/NCEAS/datateam-training/blob/master/workflows/solr_queries/construct_a_query.Rmd
}
