% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{get_unit_split}
\alias{get_unit_split}
\title{Performs formatting on unit and splits into components}
\usage{
get_unit_split(unit, all_units = mem_load_all_units())
}
\arguments{
\item{unit}{(character) unit}

\item{all_units}{(data.frame)}
}
\value{
(character) unit_split
}
\description{
Performs formatting on unit and splits into components
}
