% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_member_node.R
\name{guess_member_node}
\alias{guess_member_node}
\title{Guess the Member Node}
\usage{
guess_member_node(pid, cn = "PROD")
}
\arguments{
\item{pid}{(character) The Datone unique object identifier.  A Dataone package URL can also be used as input (although this method is less reliable).}

\item{cn}{(character) A character vector of coordinate nodes to search.  Defaults to "PROD".
Can be set to any combination of ("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2").}
}
\description{
Guess the member node that stores a Dataone object based on its
unique identifier (\code{pid}) and coordinating node (\code{cn}).  In most cases
the object is stored on the Production ("PROD") Node, however this function can
search across all coordinating nodes.  If only one member node is identified this
function returns the member node as a Dataone "MNode" object.  If multiple member
nodes are identified a vector of nodes is printed.
}
\examples{
\dontrun{
The following two calls are equivalent:
mn <- guess_member_node("doi:10.18739/A2G287")
mn <- guess_member_node("doi:10.18739/A2G287", "PROD")

Use a DataOne package URL
mn <- guess_member_node("https://search.dataone.org/#view/https://pasta.lternet.edu/package/metadata/eml/knb-lter-ntl/276/13")
Search all coordinating nodes:
cn = c("PROD", "STAGING", "STAGING2", "SANDBOX", "SANDBOX2", "DEV", "DEV2")
mn <- guess_member_node("doi:10.18739/A2G287", cn)
}

}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
