% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes_html.R
\name{hot_attributes_table}
\alias{hot_attributes_table}
\title{Handsontable attributes}
\usage{
hot_attributes_table(df, type = NULL)
}
\arguments{
\item{df}{(data.frame) the data.frame of data that needs an attribute table}

\item{type}{(character) either "attributes", "units", or "factors"}
}
\description{
Function to call handsontable widget to build attributes
}
