% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{load_EML_units}
\alias{load_EML_units}
\title{Loads EML unit library from EML::get_unitList()}
\usage{
load_EML_units(all_units = mem_load_all_units())
}
\arguments{
\item{all_units}{(data.frame)}
}
\value{
(list) EML_units
}
\description{
Loads EML unit library from EML::get_unitList()
}
