% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_category.R
\name{new_catagory}
\alias{new_catagory}
\title{Interim method of categorizing datasets}
\usage{
new_catagory(doi, themes, coder)
}
\arguments{
\item{doi}{(character) the doi}

\item{themes}{(character) themes - definition of the themes can be found here}

\item{coder}{(character)}
}
\value{

}
\description{
Interim method of categorizing datasets
}
