% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{one_package_copy}
\alias{one_package_copy}
\title{Copy a Data Package without its child packages.}
\usage{
one_package_copy(mn_pull, mn_push, resource_map_pid)
}
\arguments{
\item{mn_pull}{(MNode) The Member Node to download from.}

\item{mn_push}{(MNode) The Member Node to upload to.}

\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}
}
\value{
(list) List of all the identifiers in the new Data Package.
TODO - better name for this function?
}
\description{
The wrapper function 'package_copy' should be used instead.
This function copies a data package from one DataOne member node to another,
excluding any child data packages.
}
