% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{output_text_func}
\alias{output_text_func}
\title{Outputs data.frame to text for shiny app}
\usage{
output_text_func(df)
}
\arguments{
\item{df}{(data.frame)}
}
\description{
Outputs data.frame to text for shiny app
}
