% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{package_copy}
\alias{package_copy}
\title{Copy a Data Package}
\usage{
package_copy(mn_pull, mn_push, resource_map_pid)
}
\arguments{
\item{mn_pull}{(MNode) The Member Node to download from.}

\item{mn_push}{(MNode) The Member Node to upload to.}

\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}
}
\description{
This function copies a data package from one DataOne member node to another.
}
