% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{pid_to_prefix}
\alias{pid_to_prefix}
\title{Remove and substitute special characters in a dataOne identifier}
\usage{
pid_to_prefix(pid)
}
\arguments{
\item{pid}{(character) The identifier a dataOne object}
}
\value{
(character) The formatted identifer as a string
}
\description{
This is a helper function for the 'download_package' function. Remove special
characters from a package identifier for use as a filename prefix. Can be
generalized for use with any pid.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
