% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pkg_structure.R
\name{plot_pkg_structure}
\alias{plot_pkg_structure}
\title{Plot package structure}
\usage{
plot_pkg_structure(mn, parent_rm_pid)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{parent_rm_pid}{(character) The top-level PID in a data package family.}
}
\value{
A visIgraph plot.
}
\description{
This function visualizes how data packages in a data package family
are related to each other.
}
\examples{
\dontrun{
cn <- CNode("PROD")
mn <- getMNode(cn,"urn:node:ARCTIC")

parent_rm_pid <- "resource_map_urn:uuid:..."

plot_pkg_structure(mn, parent_rm_pid)
}
}
