% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_abstract}
\alias{qa_abstract}
\title{Check if abstract is present with sufficient length}
\usage{
qa_abstract(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if an abstract is present with sufficient length.
At least 100 words are required.
}
\details{
importFrom stringr str_split
importFrom xml2 as_list
}
\keyword{internal}
