% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_congruence}
\alias{qa_congruence}
\title{Check congruence of data and metadata attributes for all tabular data objects}
\usage{
qa_congruence(mn, resource_map_pid, read_all_data = TRUE,
  all_results = FALSE)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The PID for a resource map.}

\item{read_all_data}{(logical) Read all data from remote. If \code{FALSE}, only read first 10 rows.}

\item{all_results}{(logical) Return all results. If \code{FALSE}, only returns results with FAILURE or ERROR status.}
}
\value{
(list) A list of results.
}
\description{
This function checks the congruence of data and metadata attributes for all
tabular data objects in a data package. Supported objects include dataTable,
otherEntity, and spatialVector entities. This function also checks if all
tabular data objects can be successfully downloaded from a DataONE Member Node.
The data is downloaded based on the distribution URL found in the physical section
of the EML for supported data objects.
}
\details{
This function can be used on its own or as an option with \code{\link[=qa_package]{qa_package()}}.

This function only checks one data package at a time. For parent packages, it does not automatically check
child packages. These should be checked individually. However, it is possible to iterate through
child packages using:\preformatted{pkg <- arcticdatautils::get_package(mn, parent_rm_pid, file_names = TRUE)
results <- lapply(pkg$child_packages, function(x) qa_congruence(mn, x))
}
}
\examples{
\dontrun{
# Read all data
qa_congruence(adc_prod, "resource_map_urn:uuid:...")

# For large or many data objects, read first 10 rows
qa_congruence(adc_prod, "resource_map_urn:uuid:...", read_all_data = FALSE)

# All results
qa_congruence(adc_prod, "resource_map_urn:uuid:...", all_results = TRUE)
}
}
\seealso{
\code{\link[=qa_attributes]{qa_attributes()}} \code{\link[=qa_package]{qa_package()}}
}
