% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_contact}
\alias{qa_contact}
\title{Check if contact is present}
\usage{
qa_contact(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if a contact is present.
}
\keyword{internal}
