% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_contact_info}
\alias{qa_contact_info}
\title{Check if contact information is present}
\usage{
qa_contact_info(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if contact information is present, including an ID, email, and address.
}
\keyword{internal}
