% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_creator}
\alias{qa_creator}
\title{Check if creator is present}
\usage{
qa_creator(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if a creator is present.
}
\keyword{internal}
