% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_creator_info}
\alias{qa_creator_info}
\title{Check if creator information is present}
\usage{
qa_creator_info(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if creator information is present, including an ID, email, and address.
}
\keyword{internal}
