% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_package.R
\name{qa_creator_ORCIDs}
\alias{qa_creator_ORCIDs}
\title{Check the ORCIDs of creators in a given EML}
\usage{
qa_creator_ORCIDs(eml)
}
\arguments{
\item{eml}{(eml) Package metadata.}
}
\value{
creator_ORCIDs (character) Returns \code{character(0)} if any tests fail.
}
\description{
This function is called by \code{\link{qa_package}}.
See \code{\link{qa_package}} documentation for more details.
}
