% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_creators}
\alias{qa_creators}
\title{Test the ORCIDs of creators in a given EML}
\usage{
qa_creators(eml)
}
\arguments{
\item{eml}{(eml) Package metadata}
}
\value{
creator_ORCIDs (character). Returns \code{character(0)} if any tests fail.
}
\description{
This function is called by \code{\link{qa_package}}.
See \code{\link{qa_package}} help documentation for more details.
}
