% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_entity}
\alias{qa_entity}
\title{Check if name and description are present for data entities}
\usage{
qa_entity(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if a name and description are present for each data entity.
Additionally, entity names should be 100 characters or less.
}
\keyword{internal}
