% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_entity_dup}
\alias{qa_entity_dup}
\title{Check for duplicated data entities}
\usage{
qa_entity_dup(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if data entities have been duplicated
based on matching PIDs.
}
\keyword{internal}
