% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_geographic}
\alias{qa_geographic}
\title{Check if geographic coverage is present with description}
\usage{
qa_geographic(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if geographic coverage is present with a description.
}
\keyword{internal}
