% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_geographic_arctic}
\alias{qa_geographic_arctic}
\title{Check if geographic coverage intersects with Arctic}
\usage{
qa_geographic_arctic(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if geographic coverage intersects with the Arctic.
}
\keyword{internal}
