% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_geographic_coord}
\alias{qa_geographic_coord}
\title{Check if geographic coverage bounding coordinates are present}
\usage{
qa_geographic_coord(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if geographic coverage bounding coordinates are present.
}
\keyword{internal}
