% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_intellectualRights}
\alias{qa_intellectualRights}
\title{Check if data usage rights are present}
\usage{
qa_intellectualRights(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if data usage rights are present and are either CC-BY or CC-0.
}
\keyword{internal}
