% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_keywordSet}
\alias{qa_keywordSet}
\title{Check if keywords are present}
\usage{
qa_keywordSet(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if keywords are present. At least one
keyword is recommended.
}
\keyword{internal}
