% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_methods}
\alias{qa_methods}
\title{Check if methods are present and complete}
\usage{
qa_methods(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if methods are present and complete by checking
for method steps and sampling descriptions.
}
\keyword{internal}
