% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_package}
\alias{qa_package}
\title{Quality assurance for data package metadata}
\usage{
qa_package(mn, resource_map_pid, congruence = FALSE,
  read_all_data = TRUE, all_results = FALSE)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The PID for a resource map.}

\item{congruence}{(logical) Check congruence between data and and metadata attributes.}

\item{read_all_data}{(logical) Required if \code{congruence = TRUE}. Read all data from remote. If \code{FALSE}, only read first 10 rows.}

\item{all_results}{(logical) Return all results. If \code{FALSE}, only returns results with FAILURE or ERROR status.}
}
\value{
(list) A list of results.
}
\description{
For a single data package, this function checks the quality of the EML metadata
as well as the system metadata for all package elements, including
the resource map, metadata, and all data objects.
}
\details{
This function is a wrapper around \code{\link[=qa_eml]{qa_eml()}}, \code{\link[=qa_congruence]{qa_congruence()}}, and \code{\link[=qa_sysmeta]{qa_sysmeta()}}.
}
\examples{
\dontrun{
# Results with FAILURE or ERROR status without checking congruence
qa_package(knb_prod, "resource_map_urn:uuid:...")

# All results without checking congruence
qa_package(knb_prod, "resource_map_urn:uuid:...", all_results = TRUE)

# Results with FAILURE or ERROR status and including congruence checks
qa_package(knb_prod, "resource_map_urn:uuid:...", congruence = TRUE)
}
}
\seealso{
\code{\link[=qa_eml]{qa_eml()}} \code{\link[=qa_congruence]{qa_congruence()}} \code{\link[=qa_sysmeta]{qa_sysmeta()}}
}
