% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_physical}
\alias{qa_physical}
\title{Check if physical of data entity is present and complete}
\usage{
qa_physical(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks the physical of a data entity for presence, completeness, and mismatches.
}
\keyword{internal}
