% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_project}
\alias{qa_project}
\title{Check if project information is present and complete}
\usage{
qa_project(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if project information is present and complete.
The minimum recommended fields are title, personnel, abstract, and funding.
}
\keyword{internal}
