% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_pubDate}
\alias{qa_pubDate}
\title{Check if publication date is present with correct format}
\usage{
qa_pubDate(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if a publication date is present with the
correct format (YYYY or YYYY-MM-DD).
}
\keyword{internal}
