% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_sysmeta.R
\name{qa_sysmeta}
\alias{qa_sysmeta}
\title{Quality assurance for system metadata}
\usage{
qa_sysmeta(mn, resource_map_pid, all_results = FALSE)
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The PID for a resource map.}

\item{all_results}{(logical) Return all results. If \code{FALSE}, only returns results with FAILURE or ERROR status.}
}
\value{
(list) A list of results.
}
\description{
This function checks the quality of system metadata for all elements of a data package,
including the resource map, metadata, and all data objects.
}
\examples{
\dontrun{
# Results with FAILURE or ERROR status
qa_sysmeta(knb_prod, "resource_map_urn:uuid:...")

# All results
qa_sysmeta(knb_prod, "resource_map_urn:uuid:...", all_results = TRUE)
}
}
\seealso{
\code{\link[=qa_eml]{qa_eml()}}
}
