% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_taxonomic}
\alias{qa_taxonomic}
\title{Check if taxonomic coverage is present}
\usage{
qa_taxonomic(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if taxonomic coverage is present.
}
\keyword{internal}
