% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_temporal}
\alias{qa_temporal}
\title{Check if temporal coverage is present with correct format}
\usage{
qa_temporal(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if temporal coverage is present
and in the correct format (YYYY or YYYY-MM-DD).
}
\keyword{internal}
