% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_title}
\alias{qa_title}
\title{Check if title is present with sufficient length}
\usage{
qa_title(input)
}
\arguments{
\item{input}{(eml) An EML object.}
}
\value{
(list) A list of results.
}
\description{
This function checks if a title is present with the sufficient length.
A title should be between 7 and 20 words.
}
\keyword{internal}
