% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_view}
\alias{qa_view}
\title{View QA results in viewer}
\usage{
qa_view(input)
}
\arguments{
\item{input}{(list) A list of QA results.}
}
\description{
This is a convenience function for pretty output of QA results,
using \code{\link[listviewer:jsonedit]{listviewer::jsonedit()}}.
}
\examples{
\dontrun{
qa_view(qa_package(knb_prod, "resource_map_urn:uuid:..."))
}
}
\seealso{
\code{\link[=qa_package]{qa_package()}} \code{\link[=qa_eml]{qa_eml()}} \code{\link[=qa_sysmeta]{qa_sysmeta()}}
}
