% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_all_versions.R
\name{query_all_versions}
\alias{query_all_versions}
\title{Solr query all versions of a PID}
\usage{
query_all_versions(node, object_pid, fields = "*")
}
\arguments{
\item{node}{(MNode) Specify the node where the object should be searched for.}

\item{object_pid}{(character) PID for the object that you want to return information about.}

\item{fields}{(character) List of fields that you want returned in the data frame. Default
returns all non NULL fields.}
}
\value{
(data.frame) Data frame with rows for each version of the PID and columns with each query field
}
\description{
This function uses a combination of get_all_versions and solr query to return the query fields
for all versions of the specified PID. Each row of the resulting data frame corresponds to a version
and the columns are the query fields.
}
\examples{
\dontrun{
cn <- dataone::CNode("PROD")
mn <- dataone::getMNode(cn, "urn:node:ARCTIC")
df <- query_all_versions(mn, "doi:10.18739/A27D2Q670", c("id", "title", "origin", "submitter"))
View(df)
}
}
\author{
Sharis Ochs, \email{sharisnochs@gmail.com}
}
