% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_all_versions.R
\name{query_solr_metadata}
\alias{query_solr_metadata}
\title{Create data frame with Solr fields}
\usage{
query_solr_metadata(node, object_pid, fields = "*")
}
\arguments{
\item{node}{(MNode) Specify the node where the object should be searched for.}

\item{object_pid}{(character) PID for the object that you want to return information about.}

\item{fields}{(character) List of fields that you want returned in the data frame. Default of "*"
returns all non NULL fields.}
}
\value{
(data.frame) One row data frame with query fields as columns.
}
\description{
This is a helper function for 'query_all_versions' function. It simplifies solr queries
and creates a one row data frame with the specified Solr fields as the columns.
}
\author{
Sharis Ochs, \email{sharisnochs@gmail.com}
}
