% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pkg_structure.R
\name{query_tree}
\alias{query_tree}
\title{Query tree}
\usage{
query_tree(mn, parent_rm_pid)
}
\arguments{
\item{mn}{The member node}

\item{parent_rm_pid}{(character) The top-level PID to use for the query}
}
\description{
This is a helper function for \code{\link{plot_pkg_structure}}
that allows you to start with a (grand)parent PID and recursively
run Solr queries. The function outputs a tibble with
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
mn <- dataone::getMNode(cn,'urn:node:ARCTIC')

parent_rm_pid <- "urn:uuid:b3dc11f5-95e8-4f30-bef9-82464398bc5f"

query_tree(mn, parent_rm_pid)
}
}
