% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{return_eml_units}
\alias{return_eml_units}
\title{Return EML units}
\usage{
return_eml_units(units, quiet = FALSE)
}
\arguments{
\item{units}{(character) unit or vector of units}

\item{quiet}{(logical) if true will quiet console text}
}
\value{
(data.frame) custom unit data frame (will return a row of NAs if a unit cannot be formatted in an EML form)
}
\description{
Uses the udunits2 unit library to format inputted unit into an EML unit form (see examples).
}
\examples{
\dontrun{
#The following all return the same data frame.
return_eml_units('kilometersPerSquareSecond')
return_eml_units('Kilometers per seconds squared')
return_eml_units('km/s^2')
return_eml_units('km s-2')
return_eml_units('s-2 /     kilometers-1') # this works but is not advised
}
}
