% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{rt_login_interactive}
\alias{rt_login_interactive}
\title{Log into RT}
\usage{
rt_login_interactive(username = readline("Enter username: "),
  password = getPass())
}
\arguments{
\item{username}{RT username (enter in when prompted)}

\item{password}{RT password (enter in when prompted)}
}
\description{
Helper function to log into RT when downloading data. When run,
it prompts for the user's username and password.
}
\examples{
\dontrun{
rt_login_interactive()
}

}
\author{
Irene Steves
}
