% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_custom_units.R
\name{set_custom_UDUNITS}
\alias{set_custom_UDUNITS}
\title{Will return TRUE if properly loaded or FALSE if not}
\usage{
set_custom_UDUNITS()
}
\value{
(logical)
}
\description{
Will return TRUE if properly loaded or FALSE if not
}
