% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{shiny_attributes_table}
\alias{shiny_attributes_table}
\title{Build shiny UI for editing attributes table within function create_attributes_table()}
\usage{
shiny_attributes_table(att_table, data)
}
\arguments{
\item{att_table}{an attribute table built from create_attributes_table()}
}
\description{
Build shiny UI for editing attributes table within function create_attributes_table()
}
