% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{solr_package_copy}
\alias{solr_package_copy}
\title{Return the formatId of each data object in a package}
\usage{
solr_package_copy(node, package_identifier, formatType = "DATA")
}
\arguments{
\item{node}{(MNode/CNode) The Node to query for Object sizes}

\item{formatType}{(character) Optional. Filter to just Objects of the given
formatType. One of METADATA, RESOURCE, or DATA or * for all types}

\item{resource_map_pid}{(character) The identifier of the Data Package's Resource Map}
}
\value{
(character) The formatId, fileName, and identifier of each data object in a package.
TODO better name for this function?
}
\description{
Return the formatId of each data object in a package
}
