% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_eml_units.R
\name{try_units_deparse}
\alias{try_units_deparse}
\title{Will attempt to deparse unit with  units::deparse_unit(units::as.units(unit))}
\usage{
try_units_deparse(unit, exponents, exponents_numeric,
  all_units = load_all_units())
}
\arguments{
\item{unit}{(character)}

\item{exponents}{(character)}

\item{exponents_numeric}{(character)}

\item{all_units}{(data.frame)}
}
\value{
(character) unit. Fails if cannot deparse.
}
\description{
Will attempt to deparse unit with  units::deparse_unit(units::as.units(unit))
}
