% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_in_eml.R
\name{which_in_eml}
\alias{which_in_eml}
\title{Search through EMLs}
\usage{
which_in_eml(eml_list, element, test)
}
\arguments{
\item{eml_list}{(S4/List) an EML list object}

\item{element}{(character) element to evaluate}

\item{test}{(function/character) A function to evaluate (see examples). If test is a character, will evaluate if \code{element == test} (see example 1).}
}
\description{
This function returns locations within an EML list object that contain an instance where \code{test == TRUE}. See examples for more information.
}
\examples{
\dontrun{
# Question: Which creators have a surName "Smith"?
n <- which_in_eml(eml@dataset@creator,"surName","Smith")
# Answer: eml@dataset@creator[n]

# Question: Which dataTables have an entityName that begins with "2016"
n <- which_in_eml(eml@dataset@dataTable,"entityName", function(x) {grepl("^2016", x)})
# Answer: eml@dataset@dataTable[n]

# Question: Which attributes in dataTable[[1]] have a numberType "natural"?
n <- which_in_eml(eml@dataset@dataTable[[1]]@attributeList@attribute,"numberType","natural")
# Answer: eml@dataset@dataTable[[1]]@attributeList@attribute[n]

#' # Question: Which dataTables have at least one attribute with a numberType "natural"?
n <- which_in_eml(eml@dataset@dataTable,"numberType", function(x) {"natural" \%in\% x})
# Answer: eml@dataset@dataTable[n]
}
}
\keyword{eml}
