% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{build_factors}
\alias{build_factors}
\title{Build factor table}
\usage{
build_factors(att_table, data)
}
\arguments{
\item{att_table}{(data.frame) input attributes table}

\item{data}{(data.frame) input data}
}
\description{
Get factors
}
