% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_copy.R
\name{clone_one_package}
\alias{clone_one_package}
\title{Clone a Data Package without its child packages.}
\usage{
clone_one_package(mn_pull, mn_push, resource_map_pid)
}
\arguments{
\item{mn_pull}{(MNode) The Member Node to download from.}

\item{mn_push}{(MNode) The Member Node to upload to.}

\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}
}
\value{
(list) List of all the identifiers in the new Data Package.
TODO switch remaining for loops to applys
TODO add more messages
TODO better way to set physical in Update Metadata section?
TODO pull/push terminology could potentially be confusing. perhaps consider download/upload, from/to, source/new could be better?  I do like that they match well (both 4-letter p-words)
TODO since messages print in red (scary!), you might want to consider the crayon workaround you found. maybe it's worth having a discussion on our package 'style'?
}
\description{
The wrapper function 'clone_package' should be used instead. This function
copies a data package from one DataOne member node to another, excluding any
child data packages.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
