% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_package}
\alias{clone_package}
\title{Clone a Data Package between Member Nodes without its child packages.}
\usage{
clone_package(resource_map_pid, from, to)
}
\arguments{
\item{resource_map_pid}{(chraracter) The identifier of the Resource Map for the package to download.}

\item{from}{(MNode) The Member Node to download from.}

\item{to}{(MNode) The Member Node to upload to.}
}
\value{
(list) List of all the identifiers in the new Data Package.
TODO switch remaining for loops to applys
TODO add more messages
}
\description{
This function copies a Data Package from one DataOne member node to another.
It can also be used to copy an older version of a Data Package to the same
member node in order to restore it, provided that the old Package is then
obsoleted by the copied version.  This will not update the pids in the metadata.
}
\examples{
\dontrun{
cn_from <- dataone::CNode("PROD")
from <- dataone::getMNode(cn_from, "urn:node:ARCTIC")
cn_to <- dataone::CNode('STAGING')
to <- dataone::getMNode(cn_to,'urn:node:mnTestARCTIC')
clone_package("resource_map_doi:10.18739/A2RZ6X", from, to)
}

}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
