% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_package.R
\name{clone_package}
\alias{clone_package}
\title{Clones packages between Dataone Member Nodes.}
\usage{
clone_package(resource_map_pid, from, to, add_access_to, public = FALSE,
  clone_children = FALSE, new_pid = TRUE, new_submitter = NULL,
  change_auth_node = FALSE)
}
\arguments{
\item{resource_map_pid}{(character) Object pid}

\item{from}{(D1Client) D1Client to clone objects from.}

\item{to}{(D1Client) D1Client to clone objects to.}

\item{add_access_to}{(character, vector) Will give read, write, and changePermission access to all strings in vector. If no additional access is desired, set to \code{NULL}. Note, setting this to \code{NULL} could lead to situations where it is not possible to read, write, or changePermissions on the cloned object.}

\item{public}{(logical) Optional. Will set public read access.  Defaults to \code{FALSE}.}

\item{clone_children}{(logical) Optional. Will clone all children recursively if TRUE. Defaults to \code{FALSE}.}

\item{new_pid}{(logical) Optional. Will give the clone a new pid.  Defaults to \code{TRUE}.}

\item{new_submitter}{(character) Optional. Will change the submitter in the system metadata to given string if present.  Defaults to \code{NULL}.}

\item{change_auth_node}{(logical) Optional. Will change the authoritativeMemberNode in the system metadata to the cloned member node if TRUE.  Defaults to \code{FALSE}.}
}
\description{
This function copies a Data Package from one DataOne member node to another.
Note, the dateUploaded, obsoletes, and obsoletedBy fields in the sysmeta will be reset on the cloned object.
This will not update the information in the metadata object.  This can also be used
to restore an older version of a Package to a member node, provided that the user subsequently
obsoletes the version of the package that they used to create the clone.
}
\examples{
\dontrun{
# First set up the member nodes we're cloning between (in this example they are the same but could be different)
to <- dataone::D1Client("STAGING", "urn:node:mnTestARCTIC")
from <- dataone::D1Client("STAGING", "urn:node:mnTestARCTIC")

# Choose a package to clone (here a new one is created)
package <- arcticdatautils::create_dummy_package(to@mn)

# Clone object
cloned_package <- clone_package(resource_map_pid = package$resource_map
                          from = from,
                           to = to,
                           add_access_to = arcticdatautils:::get_token_subject(),
                           public = TRUE,
                           new_pid = TRUE)
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
