% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{create_attributes_table}
\alias{create_attributes_table}
\title{Create/Edit EML attributes}
\usage{
create_attributes_table(data = NULL, attributes_table = NULL)
}
\arguments{
\item{data}{(data.frame) the data.frame of data that needs an attribute table}

\item{attributes_table}{(data.frame) an existing attributes table for \code{data} that needs to be updated (if specified along with \code{data}, all non-empty fields will be used over any values automatically generated from \code{data})}
}
\description{
create/edit EML attributes, custom units, and factors in a dynamic setting
}
\examples{
\dontrun{
create_attributes_table(NULL, NULL)

data <- read.csv('Test.csv')
create_attributes_table(data, NULL)

attributes_table <- EML::get_attributes(eml@dataset@dataTable[[i]]@attributeList)$attributes
create_attributes_table(NULL, attributes_table)

create_attributes_table(data, attributes_table)
}
}
