% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamgmt_package.R
\docType{package}
\name{datamgmt}
\alias{datamgmt}
\alias{datamgmt-package}
\title{Data management utilities for curating, documenting, and publishing data}
\description{
The datamgmt R package supports management of data
packages on the Arctic Data Center (ADC) and State
of Alaska's Salmon and People (SASAP) data portals.
}
\seealso{
\href{https://nceas.github.io/datamgmt/}{datamgmt pkgdown website}
}
\author{
Data Science Fellows, Spring 2018: Dominic Mullen, Irene Steves, Mitchell Maier, Stephanie Freund, Emily O'Dean (NCEAS)
}
