% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{download_packages}
\alias{download_packages}
\title{Download one or multiple Data Packages}
\usage{
download_packages(mn, resource_map_pids, ...)
}
\arguments{
\item{mn}{(MNode) The Member Node to download from.}

\item{resource_map_pids}{(chraracter) The identifiers of the Resource Maps for the packages to download.}

\item{download_directory}{(character) The path of the directory to download the package to. Defaults to the current working directory.}

\item{check_download_size}{(logical) Optional.  Whether to check the total download size before continuing.  Setting this to FALSE speeds up the function, especially when the package has many elements.}

\item{download_child_packages}{(logical) Optional.  Whether to download data from child packages of the selected package.}

\item{prefix_file_names}{(logical) Optional.  Whether to prefix file names with the package metadata identifier.  This is useful when downloading files from multiple packages to one directory.}

\item{convert_excel_to_csv}{(logical) Optional. Whether to convert excel files to csv(s).  The csv files are downloaded as sheetName_excelWorkbookName.csv}

\item{check_first}{(logical) Optional. Whether to check the PIDs passed in as aruments exist on the MN before continuing. Checks that objects exist and are of the right format type. Setting this to FALSE speeds up the function, especially when the package has many elements.}
}
\description{
This function is wrapper for download_package It downloads all of the Data Objects in a Data Package
to the local filesystem.  It is particularly useful when a Data Package is too large to download using
the web interface.
}
\details{
Setting \code{check_download_size} to \code{TRUE} is recommended if you are uncertain of the total download size and want to avoid downloading very large Data Packages.

This function will also download any data objects it finds in any child Data Packages of the input data package.
If you would only like to download data from one Data Package, set \code{download_child_packages} to \code{FALSE}.
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
