% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_attribute.R
\name{edit_attribute}
\alias{edit_attribute}
\title{Edit a single attribute}
\usage{
edit_attribute(eml, dataTableNumber, attributeNumber,
  attributeName = NULL, attributeDefinition = NULL, domain = NULL,
  measurementScale = NULL, unit = NULL, numberType = NULL,
  definition = NULL, formatString = NULL, missingValueCode = NULL,
  missingValueCodeExplanation = NULL)
}
\arguments{
\item{eml}{The eml object containing the attributeList}

\item{dataTableNumber}{The number of the dataTable containing the attributeList; in other words its position in the dataTable list.}

\item{attributeNumber}{The number of the attribute of interest; in other words its position in the attributeList.}

\item{attributeName}{The new name to give to the attribute}

\item{attributeDefinition}{The new attributeDefinition to give to the attribute}

\item{domain}{The new domain to give to the attribute}

\item{measurementScale}{The new measurementScale to give to the attribute}

\item{unit}{The new unit to give to the attribute}

\item{numberType}{The new numberType to give to the attribute}

\item{definition}{The new definition (for textDomain) to give to the attribute}

\item{formatString}{The new formatString to give to the attribute}

\item{missingValueCode}{The new missing value code to give to the attribute}

\item{missingValueCodeExplanation}{The new missing value code explanation to give to the attribute}
}
\description{
This function edits the slots of a single attribute in an existing attribute table.
}
\details{
Can only be used on attributes entirely defined within the "attributes" slot of attributeList; it cannot be used to edit
the factor table of an enumeratedDomain. In cases with very large attribute lists, user may want to use
\code{\link[=which_in_eml]{which_in_eml()}} first to locate the attribute position.
}
\examples{
\dontrun{
#To change an attribute's name and add a missing value code where there hadn't been one:
new_eml <- edit_attribute(eml, 2, 8, attributeName = "new name", missingValueCode = "NA",
                          missingValueCodeExplanation = "Data unavailable")
# To change an attribute's attributeDefinition, as well as change measurementScale
# from nominal to ratio (requires also changing domain, unit, and numberType, as well
# as setting definition to NA):
new_eml <- edit_attribute(eml, 1, 2, attributeDefinition = "new definition",
domain = "numericDomain", measurementScale = "ratio", unit = "dimensionless",
numberType = "whole", definition = NA)
}
}
