% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_package.R
\name{excel_to_csv}
\alias{excel_to_csv}
\title{Convert excel workbook to multiple csv files}
\usage{
excel_to_csv(path, prefix = NULL)
}
\arguments{
\item{path}{(character) File location of the excel workbook.}

\item{prefix}{(character) Optional prefix to prepend to the file name.}
}
\value{
(invisible())
}
\description{
This is a helper function for download_package.
}
\author{
Dominic Mullen \email{dmullen17@gmail.com}
}
