% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_attributes.R
\name{get_all_attributes}
\alias{get_all_attributes}
\title{Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(url) Package}
\usage{
get_all_attributes(eml = NULL, url = NULL, node = NULL)
}
\arguments{
\item{eml}{(S4) Optional. The full eml document. Required if \code{url} is not specified.}

\item{url}{(character) Optional. The url to the Dataone Package. Required if \code{eml} is not specified.}

\item{node}{(MNode/CNode) Optional.  Defaults to Arctic Data Center Production node.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(url) Package
}
\examples{
\dontrun{
install_github("ropensci/EML)
library(EML)
library(dataone)
cn <- dataone::CNode('PROD')
node <- dataone::getMNode(cn,'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(node, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_all_attributes(eml)

attributes <- datamgmt::get_all_attributes(link = "https://arcticdata.io/catalog/#view/doi:10.18739/A23W02")

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_all_attributes(url = "https://knb.ecoinformatics.org/#view/doi:10.5063/F1639MWV",
                                 node = knb)
}
}
\author{
Dominic Mullen
}
