% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_awards.R
\name{get_awards}
\alias{get_awards}
\title{Get NSF Arctic/Polar program award information}
\usage{
get_awards(from_date = NULL, to_date = NULL, query = NULL,
  print_fields = NULL)
}
\arguments{
\item{from_date}{(character) Optional. Returns all
records with start date after specified date.
Format = "mm/dd/yyyy"}

\item{to_date}{(character) Optional. Returns all
records with start date before specified date.
Format = "mm/dd/yyyy"}

\item{query}{(character) Optional. By default, the function
searches for all awards with either "polar" or "arctic" in
the fundProgramName. Additional queries can be specified
as defined in the NSF API.
Use '&' to join multiple queries (i.e., "keyword=water&agency=NASA")}

\item{print_fields}{(character) Optional. By default, the
following fields will be returned: id, date,
startDate, expDate, fundProgramName, poName,
title, awardee, piFirstName, piLastName, piPhone, piEmail.
Additional field names can be found in the printFields description
of the [NSF API](https://www.research.gov/common/webapi/awardapisearch-v1.htm).}
}
\description{
Uses the [NSF API](https://www.research.gov/common/webapi/awardapisearch-v1.htm)
to get all records pertaining to the Arctic or Polar programs.
}
\examples{
\dontrun{
all_awards <- get_awards()
new_awards <- get_awards(from_date = "01/01/2017")
}
}
\author{
Irene Steves
}
