% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attributes.R
\name{get_eml_attributes}
\alias{get_eml_attributes}
\title{Return attribute (column) metadata from a DataONE metadata object}
\usage{
get_eml_attributes(doc)
}
\arguments{
\item{doc}{(emld) EML object.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Return attribute metadata from an EML object. This is largely a
wrapper for the function \code{\link[EML:get_attributes]{EML::get_attributes()}}.
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
mn <- dataone::getMNode(cn, 'urn:node:ARCTIC')
doc <- EML::read_eml(rawToChar(dataone::getObject(mn, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_eml_attributes(doc)

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
doc <- EML::read_eml(rawToChar(dataone::getObject(knb, "doi:10.5063/F1639MWV")))
attributes <- get_eml_attributes("doi:10.5063/F1639MWV")
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
