% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_attributes.R
\name{get_meta_attributes}
\alias{get_meta_attributes}
\title{Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(url) Package}
\usage{
get_meta_attributes(node = "ADC", metadata)
}
\arguments{
\item{node}{(MNode/CNode) The Dataone Node that stores the Metadata object, from \url{https://cn.dataone.org/cn/v2/node}}

\item{metadata}{(S4 / character) Optional. Either the full eml S4 object document or the url of the Dataone Package.
The S4 object input is a more reliable method.}
}
\value{
(list) A list of all attribute metadata from the EML in data.frame objects
}
\description{
Get all attribute from an EML document.  Document for the function \code{get_attributes}
from the EML(url) Package
}
\examples{
\dontrun{
node <- dataone::getMNode(cn, 'urn:node:ARCTIC')
eml <- EML::read_eml(rawToChar(dataone::getObject(node, "doi:10.18739/A23W02")))
attributes <- datamgmt::get_attributes_url(metadata = eml)

attributes <- datamgmt::get_attributes_url("ADC", "https://arcticdata.io/catalog/#view/doi:10.18739/A23W02")

# switch nodes
cn <- dataone::CNode('PROD')
knb <- dataone::getMNode(cn,"urn:node:KNB")
attributes <- get_attributes_url("KNB", "https://knb.ecoinformatics.org/#view/doi:10.5063/F1639MWV")
}
}
\author{
Dominic Mullen, \email{dmullen17@gmail.com}
}
