% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pkg_structure.R
\name{plot_pkg_structure}
\alias{plot_pkg_structure}
\title{Plot package structure}
\usage{
plot_pkg_structure(mn, parent_rm_pid)
}
\arguments{
\item{mn}{The Member Node.}

\item{parent_rm_pid}{(character) The top-level PID in a data package family.}
}
\description{
This function allows you to quickly visualize how data packages
are related to each other.
}
\examples{
\dontrun{
cn <- dataone::CNode('PROD')
mn <- dataone::getMNode(cn,'urn:node:ARCTIC')

parent_rm_pid <- "resource_map_urn:uuid:2b5f44fc-810e-4c30-a63f-5bd2f3ff00a7"

plot_pkg_structure(mn, parent_rm_pid)
}
}
