% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_attributes}
\alias{qa_attributes}
\title{Test congruence of attributes and data for a given dataset and dataTable}
\usage{
qa_attributes(node, dataTable, data, checkEnumeratedDomains = TRUE)
}
\arguments{
\item{node}{(MNode) Member Node where the PID is associated with an object.}

\item{dataTable}{(dataTable) EML \code{dataTable} associated with the data object.}

\item{data}{(data.frame) Data frame of data object.}

\item{checkEnumeratedDomains}{(logical) Default TRUE. If True, will match unique values in data to defined EML enumerated domains.}
}
\description{
This function is called by get_package() but can be used on its own to test congruence
between a dataTable and a data object (data.frame). See qa_package() help documentation for more details.
}
\examples{
\dontrun{
# For a package, run QA checks on a data.frame and its associated EML dataTable.

qa_attributes(mn, dataTable, dataObject)
}
}
\author{
Emily O'Dean \email{eodean10@gmail.com}
}
