% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_eml}
\alias{qa_eml}
\title{Quality assurance for EML metadata}
\usage{
qa_eml(input, all_results = FALSE)
}
\arguments{
\item{input}{(eml) An EML object.}

\item{all_results}{(logical) Return all results. If FALSE, only returns results with FAILURE status.}
}
\value{
(list) A list of results.
}
\description{
This function checks EML fields for presence, completeness, correct formatting, etc.
}
\examples{
\dontrun{
pkg <- get_package(adc_test, rm_pid, file_names = TRUE)
eml <- read_eml(getObject(adc_test, pkg$metadata))
results <- qa_eml(eml) }
}
