% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_package}
\alias{qa_package}
\title{Test package including congruence of attributes and data}
\usage{
qa_package(node, pid, readAllData = TRUE, check_attributes = TRUE,
  check_creators = FALSE, check_access = FALSE)
}
\arguments{
\item{node}{(MNode) Member Node where the PID is associated with a package.}

\item{pid}{(character) The PID of a resource map to be QA'ed.}

\item{readAllData}{(logical) Default TRUE. If True, pull all data from remote and check that column types match attributes, otherwise only pull first 10 rows. Only applicable to public packages (private packages will read complete dataset). If \code{check_attributes = FALSE}, no rows will be read.}

\item{check_attributes}{(logical). Default TRUE. Calls \code{\link{qa_attributes}}.}

\item{check_creators}{(logigal) Default FALSE. If True, will test if each creator has been assigned an ORCID. Will also run if \code{check_access = TRUE}.}

\item{check_access}{(logigal) Default FALSE. If True, will test if each creator has full access to the metadata, resource_map, and data objects. Will not run if the tests associated with \code{check_creators} fails.}
}
\value{
invisible
}
\description{
This script assumes correctness in the resource map and data
Purpose: QA script to check a DataONE EML package. Will check that attributes match values in the data.
}
\details{
Note: this function also calls \code{\link{qa_attributes}} and passes the data object and associated dataTable, but that function can also be called directly.
}
\examples{
\dontrun{
# For a package, run QA checks
qa_package(mn, pid, readAllData = FALSE, check_attributes = TRUE,
           check_creators = FALSE, check_access = FALSE)
}
}
\author{
Emily O'Dean \email{eodean10@gmail.com}
}
