% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_package.R
\name{qa_package}
\alias{qa_package}
\title{Check package including congruence of attributes and data}
\usage{
qa_package(mn, resource_map_pid, read_all_data = TRUE,
  check_attributes = TRUE, check_creators = FALSE,
  check_access = FALSE, check_attribute_classes = TRUE, skip = 0,
  delimiter = "")
}
\arguments{
\item{mn}{(MNode) The Member Node to query.}

\item{resource_map_pid}{(character) The PID for a resource map.}

\item{read_all_data}{(logical) Read all data from remote and check that column types match attributes. If \code{FALSE},
only read first 10 rows. Only applicable to public packages (private packages will read complete dataset).
If \code{check_attributes = FALSE}, no rows will be read.}

\item{check_attributes}{(logical) Check congruence of attributes and data.}

\item{check_creators}{(logical) Check if each creator has an ORCID. Will also run if \code{check_access = TRUE}.}

\item{check_access}{(logical) Check if each creator has full access to the metadata, resource map, and data objects.
Will not run if the checks associated with \code{check_creators} fail.}

\item{check_attribute_classes}{(logical) Check if column types match attribute measurementscale.  For example, a column with "ratio"
measurementScale should contain integer or numeric data.  However, these checks often fail, as read.table will read
columns in as character values - which triggers a warning.  Set \code{check_attribute_classes = FALSE} to skip
these checks.}

\item{skip}{(integer) The number of rows to skip when reading in data files.  This is useful when any metadata lines above
tabular have a different number of columns - resulting in an error reading in the file. Skip these lines but do not
skip column headers.}

\item{delimiter}{(character) the field separator character. Values on each line of the file are separated by this character.}
}
\value{
\code{NULL}
}
\description{
This function checks that the attributes listed in the metadata match the values in the data for each
tabular data object. It may also optionally check if all creators have ORCIDs and have full access
to all elements of the data package.
}
\examples{
\dontrun{
# Run all QA checks

qa_package(mn, pid, read_all_data = TRUE, check_attributes = TRUE,
           check_creators = TRUE, check_access = TRUE)
}
}
