% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_data_qa.R
\name{qa_package}
\alias{qa_package}
\title{Check package including congruence of attributes and data}
\usage{
qa_package(node, pid, readAllData = TRUE, check_attributes = TRUE,
  check_creators = FALSE, check_access = FALSE)
}
\arguments{
\item{node}{(MNode) Member Node where the PID is located.}

\item{pid}{(character) The PID of a resource map.}

\item{readAllData}{(logical) Default TRUE. Read all data from remote and check that column types match attributes, otherwise only pull first 10 rows. Only applicable to public packages (private packages will read complete dataset). If \code{check_attributes = FALSE}, no rows will be read.}

\item{check_attributes}{(logical) Default TRUE. Calls \code{\link{qa_attributes}}. Checks congruence of attributes and data.}

\item{check_creators}{(logigal) Default FALSE. Calls \code{\link{qa_creator_ORCIDs}}. Checks if each creator has an ORCID. Will also run if \code{check_access = TRUE}.}

\item{check_access}{(logigal) Default FALSE. Calls \code{\link{qa_access}}. Checks if each creator has full access to the metadata, resource map, and data objects. Will not run if the checks associated with \code{check_creators} fail.}
}
\value{
invisible
}
\description{
This function assumes correctness in the resource map and data.
Purpose: QA script to check a DataONE EML package. Checks that attributes match values in the data.
Note: this function also calls \code{\link{qa_attributes}}, \code{\link{qa_creator_ORCIDs}}, and \code{\link{qa_access}}.
}
\examples{
\dontrun{
# Run QA checks

qa_package(mn, pid, readAllData = FALSE, check_attributes = TRUE,
           check_creators = FALSE, check_access = FALSE)
}
}
