% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_eml.R
\name{qa_view}
\alias{qa_view}
\title{View QA results in viewer}
\usage{
qa_view(input)
}
\arguments{
\item{input}{(list) A list of QA results.}
}
\description{
This is a convenience function for pretty output of QA results,
using \code{\link[listviewer:jsonedit]{listviewer::jsonedit()}}.
}
\examples{
\dontrun{
qa_view(qa_eml(eml))

results <- qa_eml(eml)
qa_view(results)
}
}
\seealso{
\code{\link[=qa_eml]{qa_eml()}} \code{\link[=qa_sysmeta]{qa_sysmeta()}}
}
