% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_attributes_table.R
\name{shiny_attributes_table}
\alias{shiny_attributes_table}
\title{Create/Edit EML attributes}
\usage{
shiny_attributes_table(att_table, data)
}
\arguments{
\item{att_table}{(data.frame) initial attributes table}

\item{data}{(data.frame) the data.frame of data that needs an attribute table}
}
\description{
create/edit EML attributes, custom units, and factors in a dynamic setting
}
